+++ fabric/src/main/java/nl/enjarai/doabarrelroll/compat/modmenu/ModMenuPlugin.java
@@ -1,0 +1,12 @@
+package nl.enjarai.doabarrelroll.compat.modmenu;
+
+import nl.enjarai.cicada.api.compat.CompatMixinPlugin;
+
+import java.util.Set;
+
+public class ModMenuPlugin implements CompatMixinPlugin {
+    @Override
+    public Set<String> getRequiredMods() {
+        return Set.of("modmenu");
+    }
+}