+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-- Removed dependency on Forgified Fabric API.
+- Its a secret.
+- I sure hope this works on fabric tho.